Move RenderRoutes to services
